refactor(ui): extract chat/useWorkspace hook from App.tsx
Moves the workspace context chips state, the per-session context refresh,
switchWorkspace (pending pre-session picks included), and
applyPendingWorkspace (still invoked from the send pipeline before the first
POST /v1/responses) into a dedicated hook. Pure move, no behavior change.
